Phase2_RAG_Pipeline/requirements.txt
google-generativeai>=0.8.0
chromadb>=0.4.0
